// JeeFlatForm-MicroFrontend-/projects/jeemeet/src/app/page/JeeMeetModule/components/components/work-gov/models/JeeWorkModel.ts
export class TagsModel extends BaseModel {
    id_row: number;
    title: string;
    color: string;
    id_project_team: string;
    project_team: string;

    clear() {
        this.id_row = 0;
        this.color = '';
        this.title = '';
        this.id_project_team = '0';
        this.project_team = '';
    }
}